Anish

+++ app/Http/Controllers/RolesController.php
@@ -1,5 +1,7 @@
 <?php
+
 namespace App\Http\Controllers;
+
 
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -7,6 +9,7 @@ use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 use Illuminate\Support\Facades\DB;
+
 class RolesController extends Controller
 {
     public function index()
@@ -25,15 +28,16 @@ class RolesController extends Controller
         $role = Role::create(['name' => $request->input('name')]);
         return redirect()->route('roles.index')->with('success', 'Role created successfully.');
     }
-    public function fetchrole($id){
+    public function fetchrole($id)
+    {
         $permissions = DB::table('role_has_permissions')
-        ->select('permissions.name')
-        ->join('permissions', 'role_has_permissions.permission_id', '=', 'permissions.id')
-        ->where('role_has_permissions.role_id', $id)
-        ->get();
+            ->select('permissions.name')
+            ->join('permissions', 'role_has_permissions.permission_id', '=', 'permissions.id')
+            ->where('role_has_permissions.role_id', $id)
+            ->get();
         // dd($permissions);
-        return response()->json(['permissions' => $permissions]);   
-}
+        return response()->json(['permissions' => $permissions]);
+    }
 
     public function edit(Role $role)
     {
@@ -55,6 +59,7 @@ class RolesController extends Controller
 
     public function updatePermissions(Request $request, Role $role)
     {
+
         $permissions=$request->permissions;
         $user=Auth::user();
         // dd($request);
@@ -64,12 +69,15 @@ class RolesController extends Controller
                 'model_type' => get_class($user), // Assuming User model
                 'model_id' => $user->id,
             ]);
-            $roles=$role->givePermissionTo($permission);
+            $roles = $role->givePermissionTo($permission);
         }
 
-      
+
         return redirect()->route('roles.index')->with('success', 'Permissions updated successfully.');
     }
+
+ 
+
     public function updateAdminPermissions(Request $request, Role $role)
     {
         $permissions = $request->permissions;
@@ -96,9 +104,13 @@ class RolesController extends Controller
             return back()->with('success', 'User not found for the selected role.');
         }
     }
-    public function destroy(Role $role)
+
+   public function destroy($id){
     {
-        $role->delete();
-        return redirect()->route('roles.index')->with('success', 'Role deleted successfully.');
+        $role = Role::find($id);
+        if ($role) {
+            $role->delete();
+        }
+        return response()->json(['success' => true]);
     }
 }
